Started using re.findall()
Next step: Right now, the program returns the search term and a few
words next to it. It would be more useful if I could do the following:
1) Find the headline of the news item in which the search term appears.
(NB: see line 71 and following for some things that could help me here.)
2) Find the publication date. (The tip from @Fichtitious should help a
lot here.)
3) Find a few more context words surrounding the search term. Maybe get
the whole sentence. That would be cool.
4) Store all of the above into a dictionary.
5) Maybe prompt the user (me, I'm assuming) to enter a numerical rating
of how interesting or relevant the article was. Save the rating in the
dictionary.