Identied and work on step 1/4 and 1/2
Next step: Right now, the program returns the search
term and a few words next to it. It would be more useful
if I could do the following:
1/4) Get easier-to-read results for my search of 'books,'
which returns a really long list without any line spaces.
1/2) Search for uppercase variants of the terms in SEARCHTERMS.
1) Find the headline of the news item in which the search
term appears.
2) Find the publication date. (The tip from @Fichtitious
should help a lot here.
3) Find a few more context words surrounding the search term.
Maybe get the whole sentence. That would be cool.
4) Store all of the above into a dictionary.
5) Maybe prompt the user (me, I'm assuming) to enter a
numerical rating of how interesting or relevant the article
was. Save the rating in the dictionary.